requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.3
Jinja2==3.1.3
jsonfield==3.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
PyMySQL==1.1.0
python-decouple==3.8
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
